notifier-bot/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.6.0
asyncprawcore==2.3.0
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
prawcore==2.3.0
pycodestyle==2.9.1
requests==2.28.1
tomli==2.0.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==1.26.12
websocket-client==1.4.2
yarl==1.8.1
